Updated names. Fixed unit tests because output of  getFileURLSFromGitHubRepo function were not guaranteed to be in the same order every time.
